Fixed problem on rotation and scaling

+++ Engine/includes/tcp_server/Commands/RotateCommand.hpp
@@ -8,7 +8,7 @@
 
 class RotateCommand : public TcpCommand {
 public:
-  RotateCommand(uuid uuid, int new_x, int new_y, int new_z)
+  RotateCommand(uuid uuid, float new_x, float new_y, float new_z)
       : TcpCommand(uuid), _new_rotation(new_x, new_y, new_z) {}
   std::string execute(RT::Engine *engine) override;
   std::string undo() override;
